manifest system test install 1
