+++ app/global-settings.css
@@ -1,165 +1,166 @@
-body, html {
-  margin: 0;
-  padding: 0;
-  box-sizing: border-box;
-  font-family: Arial, sans-serif;
-  min-height: 100vh;
-  display: flex;
-  flex-direction: column;
-  background-color: var(--background-color, #f9f9f9);
-  color: var(--text-color, #333);
+body,
+html {
+    margin: 0;
+    padding: 0;
+    box-sizing: border-box;
+    font-family: Arial, sans-serif;
+    min-height: 100vh;
+    display: flex;
+    flex-direction: column;
+    background-color: var(--background-color, #f9f9f9);
+    color: var(--text-color, #333);
 }
 
 h2 {
-  text-align: center;
-  margin: 20px 0;
-  font-size: 24px;
-  color: var(--header-color, #222);
+    text-align: center;
+    margin: 20px 0;
+    font-size: 24px;
+    color: var(--header-color, #222);
 }
 
 h3 {
-  text-align: center;
-  margin: 20px 0;
-  font-size: 20px;
-  color: var(--header-color, #222);
+    text-align: center;
+    margin: 20px 0;
+    font-size: 20px;
+    color: var(--header-color, #222);
 }
 
 .setting-toggle {
-  display: flex;
-  align-items: center;
-  justify-content: space-between;
-  margin: 20px;
+    display: flex;
+    align-items: center;
+    justify-content: space-between;
+    margin: 20px;
 }
 
 .setting-item {
-  display: flex;
-  flex-direction: column;
-  margin: 20px;
+    display: flex;
+    flex-direction: column;
+    margin: 20px;
 }
 
 .setting-item button {
-  width: 100%;
-  box-sizing: border-box;
+    width: 100%;
+    box-sizing: border-box;
 }
 
 label {
-  margin-bottom: 5px;
-  font-size: 16px;
-  color: var(--label-color, #444);
+    margin-bottom: 5px;
+    font-size: 16px;
+    color: var(--label-color, #444);
 }
 
 input[type="checkbox"] {
-  width: 20px;
-  height: 20px;
+    width: 20px;
+    height: 20px;
 }
 
 select {
-  padding: 10px;
-  font-size: 16px;
-  border: 1px solid var(--text-color, #333);
-  border-radius: 5px;
-  outline: none;
-  background-color: var(--background-color, #f9f9f9);
-  color: var(--text-color, #333);
-  transition: border-color 0.3s, box-shadow 0.3s, background-color 0.3s, color 0.3s;
+    padding: 10px;
+    font-size: 16px;
+    border: 1px solid var(--text-color, #333);
+    border-radius: 5px;
+    outline: none;
+    background-color: var(--background-color, #f9f9f9);
+    color: var(--text-color, #333);
+    transition: border-color 0.3s, box-shadow 0.3s, background-color 0.3s, color 0.3s;
 }
 
 select:hover {
-  border-color: var(--focus-color, #61dafb);
-  box-shadow: 0 0 5px var(--focus-color, #61dafb);
+    border-color: var(--focus-color, #61dafb);
+    box-shadow: 0 0 5px var(--focus-color, #61dafb);
 }
 
 select:focus {
-  border-color: var(--focus-color, #61dafb);
-  box-shadow: 0 0 5px var(--focus-color, #61dafb);
+    border-color: var(--focus-color, #61dafb);
+    box-shadow: 0 0 5px var(--focus-color, #61dafb);
 }
 
 option {
-  background-color: var(--background-color, #f9f9f9);
-  color: var(--text-color, #333);
+    background-color: var(--background-color, #f9f9f9);
+    color: var(--text-color, #333);
 }
 
 .secondary-button {
-  background-color: var(--button-bg-color-secondary, #e0e0e0);
-  color: var(--button-text-color-secondary, #333);
-  border: none;
-  padding: 10px 20px;
-  font-size: 16px;
-  border-radius: 5px;
-  cursor: pointer;
-  transition: background-color 0.3s;
+    background-color: var(--button-bg-color-secondary, #e0e0e0);
+    color: var(--button-text-color-secondary, #333);
+    border: none;
+    padding: 10px 20px;
+    font-size: 16px;
+    border-radius: 5px;
+    cursor: pointer;
+    transition: background-color 0.3s;
 }
 
 .secondary-button:hover {
-  background-color: var(--button-hover-bg-color-secondary, #d6d6d6);
+    background-color: var(--button-hover-bg-color-secondary, #d6d6d6);
 }
 
 .primary-button {
-  background-color: var(--button-bg-color-primary, #61dafb);
-  color: var(--button-text-color-primary, #fff);
-  border: none;
-  padding: 10px 20px;
-  font-size: 16px;
-  border-radius: 5px;
-  cursor: pointer;
-  transition: background-color 0.3s;
+    background-color: var(--button-bg-color-primary, #61dafb);
+    color: var(--button-text-color-primary, #fff);
+    border: none;
+    padding: 10px 20px;
+    font-size: 16px;
+    border-radius: 5px;
+    cursor: pointer;
+    transition: background-color 0.3s;
 }
 
 .primary-button:hover {
-  transform: scale(1.05);
-  background-color: var(--button-hover-bg-color-primary, #21a1f1);
+    transform: scale(1.05);
+    background-color: var(--button-hover-bg-color-primary, #21a1f1);
 }
 
 .primary-button:active {
-  transform: scale(0.98);
+    transform: scale(0.98);
 }
 
 .footer {
-  position: fixed;
-  bottom: 0;
-  width: 100%;
-  background-color: var(--footer-bg-color, #282c34);
-  display: flex;
-  justify-content: center;
-  padding: 10px;
-  z-index: 1000;
+    position: fixed;
+    bottom: 0;
+    width: 100%;
+    background-color: var(--footer-bg-color, #282c34);
+    display: flex;
+    justify-content: center;
+    padding: 10px;
+    z-index: 1000;
 }
 
 .footer button {
-  margin: 0 10px;
+    margin: 0 10px;
 }
 
 [data-theme="light"] {
-  --background-color: #f9f9f9;
-  --text-color: #333;
-  --header-color: #222;
-  --label-color: #444;
-  --footer-bg-color: #282c34;
-  --button-bg-color-secondary: #e0e0e0;
-  --button-text-color-secondary: #333;
-  --button-hover-bg-color-secondary: #d6d6d6;
-  --button-bg-color-primary: #61dafb;
-  --button-text-color-primary: #fff;
-  --button-hover-bg-color-primary: #21a1f1;
-  --focus-color: #61dafb;
+    --background-color: #f9f9f9;
+    --text-color: #333;
+    --header-color: #222;
+    --label-color: #444;
+    --footer-bg-color: #282c34;
+    --button-bg-color-secondary: #e0e0e0;
+    --button-text-color-secondary: #333;
+    --button-hover-bg-color-secondary: #d6d6d6;
+    --button-bg-color-primary: #61dafb;
+    --button-text-color-primary: #fff;
+    --button-hover-bg-color-primary: #21a1f1;
+    --focus-color: #61dafb;
 }
 
 [data-theme="dark"] {
-  --background-color: #181a1b;
-  --text-color: #eaeaea;
-  --header-color: #ddd;
-  --label-color: #bbb;
-  --footer-bg-color: #1f2225;
-  --button-bg-color-secondary: #444;
-  --button-text-color-secondary: #eaeaea;
-  --button-hover-bg-color-secondary: #555;
-  --button-bg-color-primary: #bb86fc;
-  --button-text-color-primary: #000;
-  --button-hover-bg-color-primary: #9a67ea;
-  --focus-color: #bb86fc;
+    --background-color: #181a1b;
+    --text-color: #eaeaea;
+    --header-color: #ddd;
+    --label-color: #bbb;
+    --footer-bg-color: #1f2225;
+    --button-bg-color-secondary: #444;
+    --button-text-color-secondary: #eaeaea;
+    --button-hover-bg-color-secondary: #555;
+    --button-bg-color-primary: #bb86fc;
+    --button-text-color-primary: #000;
+    --button-hover-bg-color-primary: #9a67ea;
+    --focus-color: #bb86fc;
 }
 
 body {
-  overflow-y: auto;
-  padding-bottom: 60px;
+    overflow-y: auto;
+    padding-bottom: 60px;
 }
