+++ app/module-manager.css
@@ -11,93 +11,63 @@ html {
     color: var(--text-color, #333);
 }
 
-h2 {
-    text-align: center;
-    margin: 20px 0;
-    font-size: 24px;
-    color: var(--header-color, #222);
-}
-
+/* Headings */
+h1,
+h2,
 h3 {
     text-align: center;
     margin: 20px 0;
-    font-size: 20px;
     color: var(--header-color, #222);
 }
-
-.form-toggle {
-    display: flex;
-    align-items: center;
-    justify-content: space-between;
-    margin: 20px;
+h1 {
+    font-size: 28px;
+}
+h2 {
+    font-size: 24px;
+}
+h3 {
+    font-size: 20px;
 }
 
-input[type="checkbox"] {
-    width: 20px;
-    height: 20px;
-}
-
-.form-item {
-    display: flex;
-    flex-direction: column;
-    margin: 20px;
-}
-
-label {
-    margin-bottom: 5px;
-    font-size: 16px;
-    color: var(--label-color, #444);
-}
-
-input[type="text"] {
-    padding: 10px;
-    font-size: 16px;
-    border: 1px solid #ccc;
-    border-radius: 5px;
-    outline: none;
-    transition: border-color 0.3s, background-color 0.3s, color 0.3s;
-    background-color: var(--input-bg-color, #fff);
-    color: var(--input-text-color, #000);
-}
-
-input[type="text"]:focus {
-    border-color: var(--focus-color, #61dafb);
-}
-
-select {
-    padding: 10px;
-    font-size: 16px;
-    border: 1px solid var(--text-color, #333);
-    border-radius: 5px;
-    outline: none;
-    background-color: var(--background-color, #f9f9f9);
-    color: var(--text-color, #333);
-    transition: border-color 0.3s, box-shadow 0.3s, background-color 0.3s, color 0.3s;
-}
-
-select:hover {
-    border-color: var(--focus-color, #61dafb);
-    box-shadow: 0 0 5px var(--focus-color, #61dafb);
-}
-
-select:focus {
-    border-color: var(--focus-color, #61dafb);
-    box-shadow: 0 0 5px var(--focus-color, #61dafb);
-}
-
-option {
-    background-color: var(--background-color, #f9f9f9);
-    color: var(--text-color, #333);
-}
-
-.selector {
-    display: flex;
-    align-items: center;
-}
-
-.selector input {
+/* Container */
+#modules-container {
     flex: 1;
-    margin-right: 10px;
+    padding: 0 20px;
+}
+
+/* Table styling */
+table {
+    width: 100%;
+    border-collapse: collapse;
+    margin: 20px auto;
+    box-shadow: 0 2px 5px rgba(0, 0, 0, 0.1);
+    background-color: var(--table-bg-color, #fff);
+}
+thead {
+    background-color: var(--header-bg-color, #eaeaea);
+}
+th,
+td {
+    padding: 12px 15px;
+    text-align: left;
+    border-bottom: 1px solid #ddd;
+    color: var(--text-color, #333);
+}
+th {
+    font-weight: 600;
+}
+tbody tr:hover {
+    background-color: var(--row-hover-bg, #f5f5f5);
+}
+
+/* Buttons */
+button {
+    font-size: 14px;
+    padding: 8px 12px;
+    border: none;
+    border-radius: 4px;
+    cursor: pointer;
+    transition: background-color 0.3s, transform 0.1s;
 }
 
 .secondary-button {
@@ -150,12 +120,32 @@ option {
     margin: 0 10px;
 }
 
+/* Inputs & selects (if added later) */
+input[type="text"],
+select {
+    padding: 10px;
+    font-size: 16px;
+    border: 1px solid #ccc;
+    border-radius: 5px;
+    outline: none;
+    transition: border-color 0.3s, background-color 0.3s, color 0.3s;
+    background-color: var(--input-bg-color, #fff);
+    color: var(--input-text-color, #000);
+}
+input[type="text"]:focus,
+select:focus {
+    border-color: var(--focus-color, #61dafb);
+    box-shadow: 0 0 5px var(--focus-color, #61dafb);
+}
+
 [data-theme="light"] {
     --background-color: #f9f9f9;
     --text-color: #333;
     --header-color: #222;
     --label-color: #444;
-    --focus-color: #61dafb;
+    --table-bg-color: #fff;
+    --header-bg-color: #eaeaea;
+    --row-hover-bg: #f5f5f5;
     --footer-bg-color: #282c34;
     --button-bg-color-secondary: #e0e0e0;
     --button-text-color-secondary: #333;
@@ -163,8 +153,7 @@ option {
     --button-bg-color-primary: #61dafb;
     --button-text-color-primary: #fff;
     --button-hover-bg-color-primary: #21a1f1;
-    --input-bg-color: #fff;
-    --input-text-color: #000;
+    --focus-color: #61dafb;
 }
 
 [data-theme="dark"] {
@@ -172,19 +161,20 @@ option {
     --text-color: #eaeaea;
     --header-color: #ddd;
     --label-color: #bbb;
-    --focus-color: #bb86fc;
     --footer-bg-color: #1f2225;
+    --header-bg-color: #2a2d31;
+    --row-hover-bg: #2e3135;
+    --table-bg-color: #1f2225;
     --button-bg-color-secondary: #444;
     --button-text-color-secondary: #eaeaea;
     --button-hover-bg-color-secondary: #555;
     --button-bg-color-primary: #bb86fc;
     --button-text-color-primary: #000;
     --button-hover-bg-color-primary: #9a67ea;
-    --input-bg-color: #1f2225;
-    --input-text-color: #eaeaea;
+    --focus-color: #bb86fc;
 }
 
 body {
     overflow-y: auto;
-    padding-bottom: 60px; /* big foot(er) :) */
+    padding-bottom: 60px;
 }
