+++ modules/mkxpz/manifest.json
@@ -1,5 +1,5 @@
 {
-    "name": "mkxpz",
+    "name": "MKXP-Z",
     "version": "1.0.0",
     "description": "A module to launch MKXP-Z native games like RPG Maker XP, VX, and VX Ace.",
     "author": "m5kro",
